# openapi.yaml
openapi: 3.1.0
info:
  title: OpenPerpus API
  version: 0.1.0
  description: CMS Perpustakaan — 35 route handler (30 resource + health/readyz/docs/metrics/register). Deploy utama Vercel (ADR-002), Cloudflare Workers alternatif; preview lokal via `npm run dev` / `npm run cf:preview`. Enforced bidirectionally by tests/openapi-conformance.test.ts.
servers:
  - url: http://localhost:3000
    description: local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token (JWT) from client-side auth; browser sessions use cookies equivalently
  schemas:
    ErrorResponse:
      type: object
      description: Error envelope emitted by jsonError (src/lib/http-error.ts). details is a SIBLING of error, never nested inside it.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                [
                  VALIDATION,
                  NOT_FOUND,
                  CONFLICT,
                  UNAUTHORIZED,
                  FORBIDDEN,
                  THEME_FORBIDDEN,
                  HINT_USE_THEME_ENDPOINT,
                  INTERNAL,
                  SAVE_FAILED,
                  FETCH_FAILED,
                  DELETE_FAILED,
                  INVALID_JSON,
                  BAD_REQUEST,
                  SPAM_DETECTED,
                  RATE_LIMITED,
                  GONE,
                  CSRF_MISMATCH,
                  MEMBER_HAS_FINES,
                  MEMBER_OVERDUE,
                  MEMBER_LOAN_LIMIT,
                  MEMBER_CHECK_FAILED,
                ]
            message:
              type: string
        details:
          description: Optional machine-readable detail payload (sibling of error)
      required: [error]
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Items per page (articles/banners/books/members/loans default 10; other collections default 20)
paths:
...
  /api/menus:
    get:
      summary: List menus (public; ?all=1 staff elevation)
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: ok
    post:
      summary: Create menu (staff)
      security:
        - BearerAuth: []
      responses:
        '201':
          description: created
  /api/menus/{id}:
    get:
      summary: Get menu by id (public when active; staff otherwise)
      security:
        - BearerAuth: []
        - {}
      responses:
        '200':
          description: ok
    put:
      summary: Update menu (staff)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ok
    delete:
      summary: Delete menu (admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ok
...
  /api/reservations:
    get:
      summary: List reservations (auth required; members scoped to own)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: ok
        '401':
          description: Silakan login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create reservation (auth required)
      security:
        - BearerAuth: []
      responses:
        '201':
          description: created
        '401':
          description: Silakan login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'Anggota tidak layak checkout: MEMBER_HAS_FINES (tagihan belum lunas), MEMBER_OVERDUE (pinjaman terlambat), MEMBER_LOAN_LIMIT (batas pinjaman aktif), atau CONFLICT (reservasi pending ganda)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Buku sudah tidak aktif (GONE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/services/{id}:
    get:
      summary: Get service by id (public when active; staff otherwise)
      security:
        - BearerAuth: []
        - {}
      responses:
        '200':
          description: ok
    put:
      summary: Update service (staff)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ok
    delete:
      summary: Delete service (admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ok
...